requirements.txt
streamlit
google-generativeai
scikit-learn
beautifulsoup4
requests
numpy